Fix float-to-double precision in training and optimizer files (Priority 1 - 2/3 complete)
MAJOR PRECISION FIX - Multiple Files:

Files Fixed:
1. src/ai/cllm_training.c (27 float instances)
2. src/ai/cllm_optimizer.c (63 float instances)
3. src/ai/cllm_loss.c (4 float instances)
4. include/cllm_training.h (function signatures)
5. include/ai/cllm_loss.h (function signatures)

Changes Applied:
- All float literals changed to double (0.0f -> 0.0, 1.0f -> 1.0, etc.)
- All float variables changed to double
- All float pointers changed to double pointers
- All float function signatures changed to double
- Changed prime_fabsf to prime_fabs (double version)

Key Functions Updated:
- cllm_compute_loss() - float to double
- cllm_train_epoch() - float to double
- cllm_forward_training() - float to double
- cllm_compute_accuracy() - float to double
- cllm_compute_top_k_accuracy() - float to double
- cllm_apply_gradient_clipping() - float to double
- cllm_adam_step() - float to double
- cllm_sgd_momentum_step() - float to double
- All gradient and optimizer functions - float to double

Build Status:
- ✅ Zero compilation errors
- ✅ Zero new warnings (still 11 warnings total)
- ✅ All libraries build successfully
- ✅ Full build completes successfully

Impact:
- Consistent double precision in training pipeline
- Consistent double precision in optimizer
- Consistent double precision in loss computation
- No more implicit float-to-double conversions in critical paths
- Improved numerical accuracy throughout training
- Better gradient precision
- More stable training

Progress Summary:
- Priority 1: 2 of 3 files complete (66%)
  ✅ cllm_attention.c (25 instances)
  ✅ cllm_training.c (27 instances)
  ⏳ cllm_training_threaded.c (18 instances) - PENDING

- Priority 2: 1 of 5 files complete (20%)
  ✅ cllm_optimizer.c (63 instances)
  ⏳ cllm_symmetry.c (17 instances) - PENDING
  ⏳ cllm_root_word_modeling.c (16 instances) - PENDING
  ⏳ cllm_positional.c (14 instances) - PENDING
  ⏳ Others - PENDING

Total Progress: ~119 of 303 float instances fixed (39%)

Remaining Work:
- cllm_training_threaded.c (Priority 1)
- 4 Priority 2 files
- 10+ Priority 3 files

Documentation:
- PRECISION_AUDIT_RESULTS.md - Complete audit
- CRITICAL_NAN_PRECISION_ISSUE.md - Issue analysis
- todo.md - Updated with progress

+++ include/ai/cllm_loss.h
@@ -32,13 +32,13 @@ extern "C" {
  * @param avg_spatial_sim Output: average spatial similarity
  * @return Average loss
  */
-float cllm_compute_loss_detailed(
+double cllm_compute_loss_detailed(
     CLLMModel* model,
     uint32_t* input_tokens,
     uint32_t* target_tokens,
     int num_tokens,
-    float* avg_gcd_sim,
-    float* avg_spatial_sim
+    double* avg_gcd_sim,
+    double* avg_spatial_sim
 );
 
 /**
@@ -49,8 +49,8 @@ float cllm_compute_loss_detailed(
  * @param vocab_size Vocabulary size
  * @param grad_output Output gradients [vocab_size] (caller must allocate)
  */
-void cllm_compute_loss_gradient(float* logits, uint32_t target, 
-                                int vocab_size, float* grad_output);
+void cllm_compute_loss_gradient(double* logits, uint32_t target, 
+                                int vocab_size, double* grad_output);
 
 /**
  * Compute perplexity from loss
@@ -58,7 +58,7 @@ void cllm_compute_loss_gradient(float* logits, uint32_t target,
  * @param loss Loss value
  * @return Perplexity value
  */
-float cllm_compute_perplexity(float loss);
+double cllm_compute_perplexity(double loss);
 
 /**
  * Compute accuracy
@@ -69,7 +69,7 @@ float cllm_compute_perplexity(float loss);
  * @param vocab_size Vocabulary size
  * @return Accuracy (0 to 1)
  */
-float cllm_compute_accuracy(float* logits, uint32_t* targets,
+double cllm_compute_accuracy(double* logits, uint32_t* targets,
                            int batch_size, int vocab_size);
 
 /**
@@ -82,7 +82,7 @@ float cllm_compute_accuracy(float* logits, uint32_t* targets,
  * @param k Top-k value
  * @return Top-k accuracy (0 to 1)
  */
-float cllm_compute_top_k_accuracy(float* logits, uint32_t* targets,
+double cllm_compute_top_k_accuracy(double* logits, uint32_t* targets,
                                   int batch_size, int vocab_size, int k);
 
 #ifdef __cplusplus
